Added home folder in Pages folder, moved home.js + home-page-btns.js to it. Deleted constants folder
